New Braze servers added
Added in the info from braze for new servers

note: syntax highlighting for "EU" string started to fail for me once there was a comma after the EU line. I left it how it was before, but still not highlighting so I am thinking I have a bad quote or something?